LPD-31885 Moves common methods to base class
